Refactor states and transitions
