+++ src/pyfoma/__init__.py
@@ -1,7 +1,6 @@
 from pyfoma.fst import FST, reverse, invert, minimize, determinize, ignore, rewrite, context_restrict, project, concatenate
 from pyfoma.paradigm import Paradigm
-from pyfoma.state import State
-from pyfoma.transition import Transition
+from pyfoma.atomic import State, Transition
 
 __author__     = "Mans Hulden"
 __copyright__  = "Copyright 2022"
